modularizado el back, 90% del put realizado
