+++ api/src/middlewares/getGameById.js
@@ -1,0 +1,15 @@
+const { Videogame, Genre } = require('../db')
+const getGameById = async (id) => {
+    let game = await Videogame.findByPk(id, {
+        include: [{
+            model: Genre,
+            attributes: ['name'],
+            through: {
+                attributes: []
+            }
+        }]
+    })
+    return game
+}
+
+module.exports = getGameById
